updated list from album rpc
